compose: reduce startup delays due to healthcheck
Use a 30s start_period for nfd and mongodb. During the start period,
health checks are run every 5s (the default start_interval) which
provides faster reaction to restarts.

With this improved mechanism, we can also increase the regular health
check interval for nfd to 1 minute to further reduce the overhead.